fix test, refactor
