+++ tinygrad/codegen/uopgraph.py
@@ -6,7 +6,6 @@ from tinygrad.dtype import dtypes, PtrDType, ImageDType, DType
 from tinygrad.ops import UnaryOps, BinaryOps, exec_alu, UOp, NOp, UOps, UPat, PatternMatcher, END_FOR_UOP, graph_rewrite, type_verify, print_uops
 from tinygrad.helpers import DEBUG, getenv, flatten, dedup, TRANSCENDENTAL, prod, CI, all_same, partition
 from tinygrad.codegen.transcendental import xexp2, xlog2, xsin, TRANSCENDENTAL_SUPPORTED_DTYPES
-from tinygrad.renderer.cstyle import CUDARenderer
 if TYPE_CHECKING: from tinygrad.renderer import Renderer
 
 # ***** float4/image store handling *****
@@ -171,27 +170,6 @@ def threefry2x32(x: UOp, seed: UOp):
     xr = [(xr[0] + ks[i % 3]), (xr[1] + ks[(i + 1) % 3] + i + 1)]
 
   return xr[1].cast(dtypes.uint64) * 2**32 | xr[0].cast(dtypes.uint64)
-
-# ***** fp8 arithmetic *****
-
-# CUDA does not handle fp8 arithmetic natively. As a workaround, we cast to float, do arithmetic and cast back.
-dts = {dtypes.fp8_e4m3, dtypes.fp8_e5m2}
-
-def _rewrite_fp8_alu(args, res):
-  dt = dtypes.float if res.arg not in (BinaryOps.CMPLT, BinaryOps.CMPNE) else dtypes.bool
-  return UOp(UOps.ALU, dt, cast(Tuple[UOp],(arg.cast(dtypes.float) for arg in args)), res.arg).cast(res.dtype)
-def srcs(n): return tuple(UPat(name=f"x{i}", dtype=dts) for i in range(n))
-
-# NOTE: match for dtypes.bool for comparison ops
-# NOTE: CUDA does not handle e4m3 <-> e5m2 casts. Done also by casting to float
-fp8_arithmetic_rewriter = PatternMatcher([
-    (UPat(UOps.ALU, dtype=dts, name="y",src=(srcs(1))), lambda x0, y: _rewrite_fp8_alu((x0,),y)),
-    (UPat(UOps.ALU, dtype=dts.union({dtypes.bool}), name="y",src=(srcs(2))), lambda x0, x1, y: _rewrite_fp8_alu((x0, x1), y)),
-    (UPat(UOps.ALU, dtype=dts, name="y",src=(srcs(3))), lambda x0, x1, x2, y: _rewrite_fp8_alu((x0, x1, x2), y)),
-    (UPat(UOps.ALU, dtype=dts, name="y",src=(srcs(3))), lambda x0, x1, x2, y: _rewrite_fp8_alu((x0, x1, x2), y)),
-    (UPat(UOps.CAST, dtype=dtypes.fp8_e4m3, src=UPat(name="x", dtype=dtypes.fp8_e5m2)), lambda x: x.cast(dtypes.float).cast(dtypes.fp8_e4m3)),
-    (UPat(UOps.CAST, dtype=dtypes.fp8_e5m2, src=UPat(name="x", dtype=dtypes.fp8_e4m3)), lambda x: x.cast(dtypes.float).cast(dtypes.fp8_e5m2)),
-  ])
 
 # ***** main rewriter *****
 
@@ -539,7 +517,6 @@ def linearize_uop(sink_in:Union[UOp, List[UOp]], opts:Optional[Renderer]=None, s
   sink: UOp = sink_in if isinstance(sink_in, UOp) else UOp(UOps.SINK, None, tuple(sink_in))
   assert sink.op is UOps.SINK, f"sink isn't sink, it's {sink.op}"
   folder = constant_folder + transcendental_folding(tuple() if TRANSCENDENTAL >= 2 or opts is None else tuple(opts.code_for_op.keys()))\
-    + fp8_arithmetic_rewriter if isinstance(opts, CUDARenderer) else PatternMatcher([])
 
   # do graph rewrite
   acc_number = 0
